move EventSourcing.ForgettablePayloads.Persistence DI code into main EventSourcing.ForgettablePayloads.Persistence package
