+++ EventSourcing.ForgettablePayloads/Persistence/PostgreSql/Extensions.DependencyInjection.PostgreSql/EventSourcingForgettablePayloadsBuilderExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using EventSourcing.ForgettablePayloads.Abstractions.DependencyInjection;
+using EventSourcing.ForgettablePayloads.Persistence.Extensions;
 
 namespace EventSourcing.ForgettablePayloads.Extensions.DependencyInjection.Persistence.PostgreSql
 {
