+++ EventSourcing.ForgettablePayloads/Persistence/PostgreSql/Extensions.DependencyInjection.PostgreSql/EventSourcingForgettablePayloadsPersistenceBuilderExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using EventSourcing.ForgettablePayloads.Persistence;
+using EventSourcing.ForgettablePayloads.Persistence.Abstractions.DependencyInjection;
 using EventSourcing.ForgettablePayloads.Persistence.PostgreSql;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
