+++ EventSourcing.ForgettablePayloads/Persistence/PostgreSql/Extensions.DependencyInjection.PostgreSql/Extensions.DependencyInjection.PostgreSql.csproj
@@ -31,7 +31,6 @@
     </ItemGroup>
     
     <ItemGroup>
-        <ProjectReference Include="..\..\ForgettablePayloads.DependencyInjection.Persistence\ForgettablePayloads.DependencyInjection.Persistence.csproj" />
       <ProjectReference Include="..\ForgettablePayloads.Persistence.PostgreSql\ForgettablePayloads.Persistence.PostgreSql.csproj" />
     </ItemGroup>
